wip - reusable, composable traversals
This is a start at defining traversals that stay in "graph land"
as long as possible... in other words, the conversion to/from
case classes is up to the caller.

This gives us a bit more flexibility when we need to get a ref to
a specific vertex, and we can e.g. chain a traversal that looks
up a blob with the "getCanonical" traversal.

It also adds an implicit `canonicalOption` method to the `GremlinScala`
class, so you can say e.g:

`Traversals.personWithExactMatch(graph.V, author).canonicalOption`

which will return an `Option[Canonical]` if there's a matching photo
(and it has a canonical).